+++ R/provider-deepseek.R
@@ -28,6 +28,7 @@ chat_deepseek <- function(
   base_url = "https://api.deepseek.com",
   api_key = deepseek_key(),
   model = NULL,
+  params = NULL,
   seed = NULL,
   api_args = list(),
   echo = NULL,
@@ -36,10 +37,13 @@ chat_deepseek <- function(
   model <- set_default(model, "deepseek-chat")
   echo <- check_echo(echo)
 
+  params <- params %||% params()
+
   provider <- ProviderDeepSeek(
     name = "DeepSeek",
     base_url = base_url,
     model = model,
+    params = params,
     seed = seed,
     extra_args = api_args,
     api_key = api_key,
@@ -49,6 +53,23 @@ chat_deepseek <- function(
 }
 
 ProviderDeepSeek <- new_class("ProviderDeepSeek", parent = ProviderOpenAI)
+
+method(chat_params, ProviderDeepSeek) <- function(provider, params) {
+  # https://platform.deepseek.com/api-docs/api/create-chat-completion
+  standardise_params(
+    params,
+    c(
+      frequency_penalty = "frequency_penalty",
+      max_tokens = "max_tokens",
+      presence_penalty = "presence_penalty",
+      stop = "stop_sequences",
+      temperature = "temperature",
+      top_p = "top_p",
+      logprobs = "log_probs",
+      top_logprobs = "top_k"
+    )
+  )
+}
 
 method(as_json, list(ProviderDeepSeek, ContentText)) <- function(provider, x) {
   x@text