+++ R/provider-openai.R
@@ -200,8 +200,9 @@ method(chat_params, ProviderOpenAI) <- function(provider, params) {
       frequency_penalty = "frequency_penalty",
       presence_penalty = "presence_penalty",
       seed = "seed",
-      max_tokens = "max_completion_tokens",
+      max_completion_tokens = "max_tokens",
       logprobs = "log_probs",
+      top_logprobs = "top_k",
       stop = "stop_sequences"
     )
   )